requirements.txt
anyio==3.6.2
APScheduler==3.9.1.post1
asyncio==3.4.3
boto3==1.26.59
botocore==1.29.59
bottle==0.12.23
cachetools==5.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
DateTime==4.9
docopt==0.6.2
google-auth==2.15.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
pipreqs==0.4.11
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pyrogram==2.0.93
PySocks==1.7.1
pyTelegramBotAPI==4.8.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==20.0a0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
telegram==0.0.1
Telethon==1.26.1
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
yarg==0.1.9
zope.interface==5.5.2
